+++ packages/nasi-lemak-carrier-services/android/src/main/java/org/nasi/lemak/carrierservices/ESimProfileStatus.kt
@@ -1,0 +1,7 @@
+package org.nasi.lemak.carrierservices
+
+enum class ESimProfileStatus(val value: String) {
+    SUCCESS("success"),
+    FAIL("fail"),
+    UNKNOWN("unknown")
+}