+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # 1.2.0
 
-- Updated `nasi` to `1.2.0`
+- Updated `nasi` to `1.2.3`
 - `nasi-lemak-react-types`
   - Added `ReducableDispatcher`
   - Added `ReducerComponent`
   - Added RxJS utility functions
   - Added `Stable` types to represent referentially-stable functions
   - Fix bug which caused `DispatchComponent` to fire side effects in the next reduce cycle.
+  - Added `Rx.asDispatch` to convert an RxJS observable to Dispatch.
+  - Added `isType` and `isScope` functions to Action.
 - `nasi-lemak-implementation`
   - Fix bug with `Theme` where themes which are registered cannot be read.
   - Fix bug with `useAsyncLegacy` where promise gets called constantly every render.
+  - Added `useRxOperator` hook to convert an RxJS Operator to React states.
 
 # 1.1.2
 
@@ -27,4 +30,3 @@
 # 1.0.0
 
 - Initial migration from internal repo
-